requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
